move check to cloud.py
